requirements.txt
alembic
asyncpg
email-validator
fastapi
passlib==1.7.4
psycopg2-binary
pydantic
pydantic-settings
python-decouple
python-dotenv
python-jose[cryptography]
python-multipart
pytest
requests
sqlalchemy
sqlalchemy_utils
sqlmodel
uvicorn[standard]